+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/amendapplicationdetails/AmendApplicationContestedAboutToStartHandler.java
@@ -1,9 +1,11 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.handler;
+package uk.gov.hmcts.reform.finrem.caseorchestration.handler.amendapplicationdetails;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.controllers.GenericAboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackHandler;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.mapper.FinremCaseDetailsMapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
@@ -20,7 +22,7 @@ import java.util.List;
 
 @Slf4j
 @Service
-public class AmendApplicationContestedAboutToStartHandler extends FinremCallbackHandler  {
+public class AmendApplicationContestedAboutToStartHandler extends FinremCallbackHandler {
 
     private static final String MIAM_INVALID_LEGACY_EXEMPTIONS_WARNING_MESSAGE =
         "The following MIAM exemptions are no longer valid and will be removed from the case data.";
